refactor(pipeline-rdf): normalize stages/ layout and canonical pipeline.test.ts
