+++ packages/core/compute/pipeline-rdf/src/stages/index.ts
@@ -1,0 +1,7 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './extract-facts';
+export * from './index-facts';
+export * from './normalize-predicates';
